add support for variable reference distances (d0). Add linestyles to lineplots. Limit scatter plots dims to spread of references.

+++ constants.py
@@ -5,8 +5,8 @@ Location = namedtuple('Location', ['x', 'y'])
 # constants
 P0:float = -28.0
 n_p:int = 2
-d0:int = 1
-X_std:float = None # specify a constant or None to calculate it from rssi-measurements
+d0:float = 1.0      # reference distance in meters
+X_std:float = None  # specify a constant or None to calculate it from rssi-measurements
 
 sensorid_to_locate = 6
 skip_ref_ids = [4,5]
